+++ files/gitlab-cookbooks/gitlab/attributes/default.rb
@@ -542,6 +542,9 @@ default['gitlab']['gitlab-pages']['redirect_http'] = false
 default['gitlab']['gitlab-pages']['use_http2'] = true
 default['gitlab']['gitlab-pages']['dir'] = "/var/opt/gitlab/gitlab-pages"
 default['gitlab']['gitlab-pages']['log_directory'] = "/var/log/gitlab/gitlab-pages"
+default['gitlab']['gitlab-pages']['artifacts_server'] = true
+default['gitlab']['gitlab-pages']['artifacts_server_url'] = nil
+default['gitlab']['gitlab-pages']['artifacts_server_timeout'] = 10
 
 ####
 # Nginx