+++ CHANGELOG.md
@@ -7,10 +7,12 @@ omnibus-gitlab repository.
 
 - Remove unused Grit configuration settings
 - Enable profiler for jemalloc
+- Add -artifacts-server and -artifacts-server-timeout support to Omnibus
 - Added PostgreSQL support for effective_io_concurrency
 - Added PostgreSQL support for max_worker_processes and max_parallel_workers_per_gather
 - Added PostgreSQL support for log_lock_waits and deadlock_timeout
 - Added PostgreSQL support for track_io_timing
+
 
 10.0.2